refactor(Bechmarks): RHICOMPL-1538 split scopes and searching
